requirements.txt
flask==1.1.4
numpy==1.19.5
gensim==4.0.0
nltk==3.2.5
tensorflow-cpu==2.7.0
tensorflow-hub==0.12.0
tensorflow-text==2.7.3
gunicorn==20.0.4
joblib==1.1.0

